Added UserDefault Manager Class and stack factory class 1. Added a detail view for listing dates 2. Added a complete detail view group with view controller , view , data source and presenter 3. Updated base navigation controller 4. Updated unit calculation logic on user selection in newly added setting page 5. Storing user selection in user defaults

+++ CleanWeather/CleanWeather/Modules/Model/ForecastData.swift
@@ -17,7 +17,7 @@ struct ForecastData: Codable {
 // MARK: - Current
 struct Current: Codable {
     let lastUpdated: String
-    let tempC: Int
+    let tempC: Double
     let tempF: Double
     let isDay: Int
     let feelslikeC: Double
